requirements.txt
Django==1.9.4
django-filter==0.13.0
django-rest-swagger==0.3.5
djangorestframework==3.3.3
Markdown==2.6.5
PyYAML==3.11
wheel==0.24.0
